requirements.txt
streamlit==1.45.1
pandas==2.3.0
fpdf2==2.8.3